+++ CHANGELOG.md
@@ -2,6 +2,25 @@
 
 ## Unreleased
 
+## v0.2.25 2019-07-31
+
+- Add `@theme-ui/custom-properties` package #235
+
+## v0.2.24 2019-07-29
+
+- Adds preset packages #210
+
+## v0.2.23 2019-07-29
+
+- Add base colors to prism presets #249
+- Add key to element in gatsby-plugin-theme-ui #248
+- Add metadata to packages #244
+- Update docs
+
+## v0.2.22 2019-07-29
+
+- Update dependencies
+
 ## v0.2.21 2019-07-26
 
 - Add presets to `@theme-ui/prism` #231